fiksa en feil
endret "double tid" til "int tid" på linje 170. Hvis det skjer feil, så si ifra så fikser me det.

+++ src/no/hvl/dat100ptc/oppgave4/GPSComputer.java
@@ -167,7 +167,7 @@ public class GPSComputer {
 		System.out.println("=================================================");
 		
 		double distanse = totalDistance() / 1000;
-		double tid = totalTime();
+		int tid = totalTime();
 		double høydeForskjell = totalElevation();
 		double maksFart = maxSpeed() * 3.6;
 		double gjenFart = averageSpeed() * 3.6;
